requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
Flask-CORS==4.0.0
langchain==0.1.0
langchain>=0.1.0
langchain-core>=0.1.0
pydantic>=2.0.0
langchain-openai>=0.0.3
openai>=1.0.0